requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0

# Scheduling
apscheduler==3.10.4

# API / Webhooks
fastapi==0.111.0
uvicorn==0.29.0

# OpenAI
openai==1.30.1

# HubSpot
hubspot-api-client==9.0.0

# Utilities
tenacity==8.3.0        # retry logic for API calls
structlog==24.2.0      # structured logging
